+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # dprint-intellij-plugin Changelog
 
 ## [Unreleased]
+- Handle execution exceptions when running can format
+- Ensure on save action is only run when on user triggered saves
 
 ## [0.3.2]
 - Fix intermittent lock up when running format